Clean up references to UPP types, plugin/host glue.
The old gecko-sdk defines a bunch of types and macros in npupp.h,
which is no longer available in xulrunner-sdk 1.9.2.  Luckily,
all of the NPN*UPP types were typedeffed to UniversalProcPtr,
so this change just uses that type, incorporated into the
HOST_TO_PLUGIN_GLUE macro.  We also define a new CALL_NPN macro
for choosing to use the CallNPN_*Proc macros from npupp.h or
just invoking the function pointer directly.

Signed-off-by: Jean-Baptiste Kempf <jb@videolan.org>