Mozilla: use --no-xlib
It will be hard to convince all Mozilla plugins-capable browsers to
call XInitThreads() early enough, if it ever happens. In the mean time,
make sure we use only pure XCB and avoid Xlib in VLC plugins.

Note: I did not check that the web plugins itself was not using Xlib
outside the main Mozilla browser thread.

This should fix #3662. We loose GLX and PulseAudio inside Mozilla.